"Exception handling"

+++ src/main/java/com/devsuperior/dscommercem/services/ProductService.java
@@ -3,11 +3,17 @@ package com.devsuperior.dscommercem.services;
 import com.devsuperior.dscommercem.dto.ProductDTO;
 import com.devsuperior.dscommercem.entities.Product;
 import com.devsuperior.dscommercem.repositories.ProductRepository;
+import com.devsuperior.dscommercem.services.exceptions.DatabaseException;
+import com.devsuperior.dscommercem.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import tools.jackson.databind.exc.InvalidDefinitionException;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,9 +27,11 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
-        return new ProductDTO(product);
-    }
+            Product product = repository.findById(id).orElseThrow(
+                    () -> new ResourceNotFoundException("Recurso não encontrado"));
+            return new ProductDTO(product);
+        }
+
 
     @Transactional(readOnly = true)
     public Page<ProductDTO> findALL(Pageable pageable){
@@ -43,24 +51,41 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (InvalidDefinitionException e){
+           throw new  ResourceNotFoundException("Recurso não encontrado");
+
+        }
 
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
-    private void copyDtoToEntity(ProductDTO dto, Product entity) {
 
-        entity.setName(dto.getName());
-        entity.setDescription(dto.getDescription());
-        entity.setPrice(dto.getPrice());
-        entity.setImgUrl(dto.getImgUrl());
+        private void copyDtoToEntity (ProductDTO dto, Product entity){
 
+            entity.setName(dto.getName());
+            entity.setDescription(dto.getDescription());
+            entity.setPrice(dto.getPrice());
+            entity.setImgUrl(dto.getImgUrl());
+
+        }
     }
-}
